Merge pull request #4986 from RosettaCommons/roccomoretti/pp_sm_addition
Add the ability to run simple metrics directly from the PROTOCOLS section.

SimpleMetrics were slightly harder to use than Filters for reporting values, as you could not add them directly to the PROTOCOL section.

This adds the ability to add SimpleMetrics with the simple directive <Add metrics=(comma separated list)/>, which evaluates them at that point in the protocol, and then adds them to the pose to be output as scores (just like if you did a RunSimpleMetrics mover.)

By default, the metrics are labeled according to their names (the strings in the metrics= line.) You can change this with a labels=(comma separated list), which will allow you to provide particular values to be used for the output. If you use '-' (a single dash) as a label, the "normal" labeling will be used.

Examples from the integration test:

    <Add metrics="L3_fa_rep" /> -- adds the L3_fa_rep metric under the label "L3_fa_rep"
    <Add metrics="total_energy_pre,rmsd,sasa" labels="TotEnergy,my_rmsd,SaSa" /> -- adds the three metrics under the labels "TotEnergy" "my_rmsd" and "SaSa", respectively.
    <Add metrics="L3_interactions,L3_no_sol" labels="-" /> -- Adds the L3_interactions metric to the output under it's default name with custom type, etc. (which happens to be "L1_interaction_energy" in the integration test case, and adds the L3_no_sol metric under the "L3_no_sol" name.
    <Add metrics="L3_fa_rep,ss" labels="interact_rep" /> -- adds the L3_fa_rep metric under the "interact_rep" name, and the ss metric under the name "ss"

The relabeling should work with string, per-residue and composite metrics, with <Add metrics="hbond" labels="pr_hbond" /> (per residue metric) giving a number of pr_hbond_{resnum} entries, and <Add metrics="protocols" /> (composite string) giving a bunch of protocols_{label} entries.

The RunSimpleMetricsMover should still work as it has before, but potentially with a greater number of options.

Old repository SHA1: 3548d5ed6e9f36f69efe414b4e30132a5077ab6d