Merge pull request #5049 from RosettaCommons/roccmoretti/disable_ligand_dock_performance
Disable the LigandDockProtocol performance test.

This test is one of the noisiest performance test. (Probably because it's attempting to do too much in each cycle.)
Additionally, it's testing an old, outdated protocol.

If we want to re-enable this, we probably want to do more granular testing (different tests for each step),
and test the most recent version of the protocol.

Old repository SHA1: 9fdd492f0eed2d37648a5b19cf9292ea64b74d65