Merge pull request #5070 from RosettaCommons/vmullig/remove_metal_connections_mover
Add a mover that removes connections that were added by the -auto_setup_metals flag or the SetupMetalsMover

It can be convenient to be able to remove bonds between metal-liganding residues and metals (_e.g._ for mutagenesis protocols in which metal-coordinating positions are mutated).  This pull request aims to add a mover that does this.  The mover will:

- Remove the bonds between liganding residues and metals.
- Remove the custom variant types from the residues.
- Remove the custom variant types from the metals.
- If a residue selector is provided, this behaviour will only apply when a liganding residue _and_ the metal are selected by the selector.  Otherwise, it will apply to the whole pose.

Tasks:
- [x] Add the mover.
     - [x] parse_my_tag()
     - [x] provide_xml_schema()
     - [x] apply()
- [x] Unit tests.
     - [x] For applying to whole pose.
     - [x] For applying to selection.
- [x] Documentation.
- [x] Pull request #4980 should be merged before this one.
- [x] This PR should be updated with changes from PR #4980 (simplified interface for CitationManager).
- [x] Beauty.

Old repository SHA1: 6de9133ab53423d54bcc1cb79b834bd4a6c923b4