Merge pull request #5079 from RosettaCommons/vmullig/suppress_unnecessary_warning
Suppress a warning message in a context in which the thing that it warns about is not an issue.

This PR suppresses warnings about out-of-date residue connections in a context in which we're about to update the residue connections anyways.  Users have raised questions about this unnecessary warning in the past, thinking that it indicates something going wrong when it does not.  This is expected to cause cosmetic changes to a bunch of integration tests, so I'm pulling it out into its own pull request.

Old repository SHA1: 6e444740edf67e0b1d3f55da70337dab122c9988