Merge pull request #5103 from RosettaCommons/vmullig/clean_up_namespaces_a_wee_bit
Clean up the namespaces a bit.

When energy methods were moved from core/scoring to core/energy_methods in order to split core.3 and create the current core.4, there was a need to finish the split quickly: linking the enormous core.3 library was taking too long and using too much memory for some people's systems, and the change needed to be in master before additional energy methods were added. Unfortunately, this meant that the energy methods retained their old namespace, breaking the convention of ensuring that directory structure and namespace match in Rosetta. This PR aims to clean this up, and to change the namespace for the energy methods that are in core/energy_methods from core::scoring to core::energy_methods.

Note: in the interests of keeping the integration tests interpretable, this PR does not change tracer names. Pull request #5108 will subsequently update the tracer names. This PR must be merged before that one.

For reviewers: Although this is a huge PR, it should only change namespaces. (The only other cases are a few places where I switched new to make_shared, and one or two places where I switched MyClass myobject = MyClass( params ) to MyClass myobject( params ).) Nothing in this PR should alter Rosetta function or output (but for possible trajectory changes that are hard to anticipate from mysterious things that the compiler does differently).

Old repository SHA1: 3eaef2aa50482bc20a543156c775ef4dd79ce8e3