Make addVirtualResAsRoot add a residue to its own chain (#5116)
I think that this is a reasonable change since it doesn't make sense to have the virtual root in the same chain grouping as the last residue of the pose that you are adding it too.

the current implementation is annoying because if you're trying to make sure the pdb_info is is line with experimental data I use pose.chain_sequence(i) to get the polymeric entities, then each are aligned. Adding an X at the end can make this annoying. If this PR breaks a lot -- then I will figure out the corner case in my own code.


* Make vrt residue on its own chain

* update ssa test

Old repository SHA1: 6c8a8f044a9959e155e55e15fd84294010a70049