Merge pull request #4987 from RosettaCommons/jadolfbr/custom_type_as_name
This PR is orthogonal to @roccomoretti 's awesome additions in PR #4986.

It adds two options to control the final naming of the metrics when used with RunSimpleMetrics.

-simple_metrics:use_name_as_custom_type
This option simply uses the tag name as the custom_type. If custom_type is given, it will override this option.

-simple_metrics:use_name_as_metric
This option uses the tag name as the core name. It ignores custom_type. Prefix/Suffix from RunSM is still added.

Old repository SHA1: 6f0b235be63a2bde515f63a0521c210958dcdbc1