Merge pull request #5186 from RosettaCommons/MentenAI/AxisCsts
This constraints the distance between an atom and the closest point to that atom on an axis. The axis is defined by two points, each of which are the midpoint of several AtomIDs. This should work with virtual atoms.

Old repository SHA1: 8c59622444c3adf4afca6e8e23db4930078ec81b