Merge pull request #5218 from RosettaCommons/vmullig/fix_glcyo_valgrind_issue
Fix an uninitialized variable issue in glycopeptide docking flagged by Valgrind

The options system read is done only if the user has specified the option, so a default value is needed for `interface_distance_` in the `GlycopeptideDockingFlags` class; otherwise this is uninitialized if the user doesn't provide a value with a commandline flag.  I'm providing a default that matches the default in both the options system and in the `RestrictToInterface` task operation (which this variable is used to configure).  Admittedly, better still would be having the default set in only one place.

Old repository SHA1: c07233a180569088c9e84814a632605cdfbcbed6