Merge pull request #5284 from RosettaCommons/roccomoretti/single_random_assert
Fix ParsedProtocol with single_random

An assert in ParsedProtocol triggers if you use the single_random protocol. The assert was erroneously testing if the number of steps was equal to the number of submovers, which isn't the case for single_random, as it only picks out a single submover.

Old repository SHA1: 02ff5238ae8d9de047217ca5bc6e82f190a1a3d7