Merge pull request #5300 from RosettaCommons/revert-5293-vmullig/shush_membrane_potential
Revert "Switching some noisy tracer output in the MembranePotential to debug output."

@jkleman informs me that currently, this is the only way to get membrane information out. In the future, we should ensure that this is written out with the pose instead of to the tracer, but for now, we'll revert this.

The discussion on the PR:

Rocco: @jkleman What's the use case here? What are you using the results thus printed for, and how often do you need it? If it's not a frequent use case, simply adding -out:levels core.scoring.MembranePotential:debug to your command line when you need it should result in things printing like they did. If it is a frequent use case, then we probably should seriously consider better ways of getting the info, other than voluminous tracer printing.

Me: I think it is a frequent use case. I agree: relying on tracer output for essential information is a bad idea. Best would be writing to the pose (on PDB output) and storing in binary silent files, but I don't have time to implement that at the moment, so temporarily I wanted to restore the old status quo.

Julia: @roccomoretti yes, it's a frequent use case. For anyone using the old RosettaMembrane code, the tracer output is the only way to get the membrane information, i.e. how the protein is embedded into the membrane. Best case would be to put it into the pose directly but fixing this hasn't been the highest priority on my TODO list. This does not affect the new RosettaMP framework.

So we'll come up with a better solution at some point, but for now we'll just make do with noisy tracers.

Old repository SHA1: 2be8bddcb7b92347a32408deb451aa1072ffac97