Merge pull request #5308 from RosettaCommons/Rosetta/trRosetta_into_Rosetta_incremental_merge_1
Incremental merge 1 for trRosetta integration into Rosetta

This PR attempts to pull the basic trRosetta calculator functionality (the `trRosettaProtocolBase` and `trRosettaProtocol_v1` classes) into their own pull request, along with all the test files for this base functionality, while leaving the mover and constraint generator (and all of their tests) for a later pull request.  It is intended to be merged before pull requests #5119 and #5309.

As a guide, the infrastructure in basic in this image is already in master.  This PR adds the stuff in protocols.2.  Subsequent PRs will add the stuff that's higher up.

![image](https://user-images.githubusercontent.com/4205776/112569688-08256000-8dbb-11eb-8cfc-2eb28fe9d3dc.png)

Old repository SHA1: 699ec9c33a981997a850931385a8ac6e7bd5811f