Merge pull request #5309 from RosettaCommons/vmullig/trRosetta_into_Rosetta_incremental_merge_2
Incremental merge 2 for putting trRosetta into Rosetta

This merge adds the trRosettaConstraintGenerator and all its associated test files.  It must be merged prior to pull request #5119.

- [x] Pull request #5308 must be merged before this one.
- [x] Need to fix a few test failures.
- [x] Andy caught a bug: a3m files that contain a comment line that starts with an octothorp cause NaN errors.  (And HHBlits produces a3m files with an initial comment line starting with an octothorp.)

As a guide, once PR #5308 is merged, the stuff in this diagram that's in protocols.2 or basic will be in master already.  This PR adds the trRosettaConstraintGenerator.  The trRosettaProtocolMover, the trRosetta application, and the modifications to the AbinitioRelax application will be in a subsequent PR.

![image](https://user-images.githubusercontent.com/4205776/112569688-08256000-8dbb-11eb-8cfc-2eb28fe9d3dc.png)

Old repository SHA1: 1c80030ac3dddd13e6d446b7ac5fda92cb370779