Merge pull request #5314 from RosettaCommons/vmullig/trRosetta_into_Rosetta_incremental_merge_3
Incremental merge #3 to get trRosetta into Rosetta

This is the third in a series of incremental merges to put trRosetta into Rosetta, to allow protocol development on top.  This PR adds the trRosettaProtocolMover, a RosettaScripts and PyRosetta-scriptable mover that takes an MSA as input and produces a predicted structure pose as output (using the trRosettaConstraintGenerator and trRosettaProtocol, already merged into master, under the hood).  The image below serves as a guide: everything below the trRosettaProtocolMover is already in master, this PR adds the trRosettaProtocolMover, and everything above it remains to be merged.

Although this PR changes 48 files, most of those are unit test files that were added.  Note that for Tensorflow modules, the unit tests are implemented as part of our integration test suite (though they are pass-fail tests), since we have no unit test suite for the Tensorflow build.

This PR is intended to be merged before PR #5119.

![image](https://user-images.githubusercontent.com/4205776/112569688-08256000-8dbb-11eb-8cfc-2eb28fe9d3dc.png)

Old repository SHA1: 711232075deb5fc323919f4fe972bff2017a6cc3