Merge pull request #5346 from RosettaCommons/vmullig/more_verbose_tensorflow_error
Make a Tensorflow error message more verbose to aid debugging

Once in a while, I'm seeing a failure in the Tensorflow tests (_e.g._ https://b3.graylab.jhu.edu/sub-test/224304352), in a piece of code that combines a vector of N-tensors into an (N+1)-tensor.  This _might_ be due to float-float comparisons (and I hope that's all it is), but I would have thought that `memcpy` would guarantee that float-float comparisons would be valid.  For this reason, I'm going to make the error messages more verbose so that we can see for ourselves what's going wrong when this test fails on the test server.

@JackMaguire  

Old repository SHA1: a81f6bd5d65d434fe237ca332eaa239883018ac9