Merge pull request #5379 from RosettaCommons/vmullig/add_ndm1_citation
Add citation for recent NDM1 paper

This PR adds a citation for doi: `10.1073/pnas.2012800118` to the Rosetta database.  It also adds citations to previously-unpublished modules (`net_charge` energy, `voids_penalty` energy, `buried_unsatisfied_penalty` energy, `hbnet` energy, `SetupMetalsMover`, `PeptideInternalHbondsMetric`) described in the supplement of that paper, and updates the citations for the simple_cycpep_predict application.

This PR also corrects a minor CitationManager issue: an extra period was being added after the article title in the citation output.  I've updated the `add_citation_by_pubmed_id.py` script to avoid adding a period to the end of the article title in the records in the database, and corrected the existing records, but this change does result in a number of minor cosmetic changes to the integration tests.

Old repository SHA1: aeb23434a504c5dd31717dda8d26efc6b26abf33