Merge pull request #5411 from RosettaCommons/vmullig/add_constraint_metric_incremental_merge_1
Incremental merge 1 for the ConstraintsMetric

To get this working with trRosetta, I need to add a way to read in and write out spline functions in a single line.  I'll continue to do that in pull request #5400, but for now the basic functionality of the `ConstraintsMetric` should work.  So that can be merged here.

This PR:
- Adds the ConstraintsMover, which summarizes the constraints in a pose or sub-region, in CST file format.
- Adds a unit test.

Before merging I'll:
- [x] Add documentation --> DONE.
- [x] Fix serialization.
- [x] Add integration test.

@jkleman

Old repository SHA1: 8ada0d888c16c4085a5e266a9673c77ed5429290