Merge pull request #5415 from RosettaCommons/BYachnin/mhc_epitope_citation_info
Update MHCEpitopeEnergy Citation Info

This PR add the citation information for the recently published MHCEpitopeEnergy paper. In addition to adding the citation to the database and registering the EnergyMethod and Constraint Mover with the CitationManager, this also adds functionality to the MHCEpitopeEnergySetup class to automatically register MHCEpitopePredictors with the CitationManager when they are used, provided that they include citation information. I have essentially tried to replicate how this is handled for Movers, etc.

Integration test failures due to added citations and changes to the database are the expected cosmetic changes.

Old repository SHA1: e8b7e0c19ac098ea29fc7340d70c2af28e4bfe25