Merge pull request #5396 from RosettaCommons/roccomoretti/cleanup_fwd
Clean up forward headers.

Part of the "try to speed up compilation" series: Go through the forward headers, and clean up unneeded includes. (Under the philosophy that the forward headers should be as minimal as possible to define the existence of types.)

A number of individual deletions, but main ones:

*    There were some "all" headers that really weren't being used, and probably shouldn't be used (include the individual subheaders instead) - those were deleted
*    There's no appreciable difference between owning_ptr.fwd.hh and owning_ptr.hh (and corresponding access_ptr files) - delete the fwd.hh versions and just use the regular hh ones.
*    Replace vector1.hh with vector1.fwd.hh in forward headers, where practical. On my machine there's about a half second of compile time difference between the two.
*    Fix up some locations where APs were being defined as shared_ptr instead of weak_ptr.

Large numbers of #include additions to fix up compilation.

Additionally, I added an option to the Windows compile on the test server to increase the number of errors it reports when it fails. (Instead of just the first.) I figure this will help when people don't have a local Windows build they can run, and have to fix the errors based on the server results.

Old repository SHA1: 16a09fec0f32092e7310388a785c370afc46d59f