Merge pull request #5484 from RosettaCommons/roccomoretti/less_python_verify
Reduce the number of times Python verification happens in integration tests.

The changes in PR #5399 exposed the fact that generateIntegrationTestGlobalSubstitutionParameters() was being called multiple times per test suite. This is unneeded, as we've already generated the parameters and have stored them in a variable we could be re-using.

Old repository SHA1: 0acecbfbee92fb1f9e499240b3c5c4661f0357d0