Merge pull request #4801 from RosettaCommons/JWLabonte/PDB_IO/splittable_residues
PDB IO: Adding a way for .pdb residues to be "split" into multiple Rosetta residues upon import

Sometimes the PDB has a 3-letter code for a ligand that Rosetta represents as an oligomer of smaller residues. For example, the small molecule maltose exists in the PDB both as the single ligand MAL and as a dimer of GLC (→4)-α-D-Glcp in Rosetta) residues.

This merge provides a way in the database to provide instructions to Rosetta to be able to split such residues properly automatically.

Old repository SHA1: 1adafb9745c9f1bad91ebce6bbc811e15d49d11f