Merge pull request #5545 from RosettaCommons/roccomoretti/fix_empty
Fix issue with empty()

For vectors, empty() is a reporting method, not an altering method. Compiling with GCC 9.1 and C++17 mode shows a number of cases where a pointless empty() is being called.

Most of these are truly pointless (the vector is definitely empty at that point, being freshly created), but there is one location where I'm not sure if the intent was to clear the vector. (Which is complicated by it being a const method.) Commenting it out preserves the current behavior, but leaves it there as a nudge to more knowledgeable people working on the code in the future.

Old repository SHA1: 7d39598934fd4a5f68c14b079a226eec4f806978