Merge pull request #5596 from RosettaCommons/vmullig/fix_single_residue_helix_error
Fix spurious error in helical_bundle_predict when PsiPred predicts a single isolated residue of helix.

The `helical_bundle_predict` application was being a bit too strict in its checks when new helices were defined.  One check was producing an error if the last residue index of the helix was not strictly greater than the first residue index.  This is unnecessarily strict since the rest of the code tolerates one-residue helices just fine: if extension beyond the defined helices is allowed, then nucleation at a single residue of helix produces one turn of helix (four helical residues), and if extension is _not_ allowed, then nucleation fails but the program continues on happily.  Since PsiPred sometimes predicts a single residue to be helical, this check was creating spurious errors in a lot of structure prediction attempts.

This pull request relaxes the check, so that now it checks that the end residue index is greater than _or equal to_ the start residue index in a helix.

This also updates the documentation submodule to point to the latest master of the documentation repo.

Old repository SHA1: 28d4aa0ee00045498e821089754a7f8376f0be7e