Merge pull request #5184 from RosettaCommons/aleaverfay/rs_schema_validator_singleton2
This commit adds a new singleton to construct the schema for RosettaScripts and to
then hold the (expensive to construct) XMLSchemaValidator. This validator can
then be requested repeatedly without requiring the costly construction of the schema
or the costly construction of the validator. This should make repeated calls
to the XMLObjects' methods much faster.

Thanks @weitzner for hashing out this idea with me. Thanks @bkuhlman for bringing
the problem to our attention.

This is the alternative fix to the issue that PR #5179 was meant to address

Old repository SHA1: 1cdf90890c5430305d08adbcb7ed2b87528e1b7c