Merge pull request #5684 from RosettaCommons/roccomoretti/update_ccd
Update the CCD components definitions

On Slack, it was mentioned that the CCD definitions in the database are about 2 years out of date at this point, and that's causing issue with the read-all-the-PDB tests.

Thanks to the scripts, it's relatively easy to update the components definitions, so we have. We've also updated the PDB_diagnostic test metadata to account for the new components.

Old repository SHA1: 9f3f54c5e1978950c38421fe54b1ad6041f27743

Co-authored-by: Steven Lewis <smlewi@gmail.com>